requirements.txt
astroid==1.5.3
backports.functools-lru-cache==1.4
click==6.7
configparser==3.5.0
enum34==1.1.6
Flask==0.12.2
isort==4.2.15
itsdangerous==0.24
Jinja2==2.9.6
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
nose==1.3.7
pycodestyle==2.3.1
pyflakes==1.5.0
pylint==1.7.4
redis==2.10.6
singledispatch==3.4.0.3
six==1.11.0
Werkzeug==0.12.2
wrapt==1.10.11
mysql-connector==2.1.4
